Unified coalesce tests
